implements test cases for chromosome and region
